Refactoring Part 2
Removed the Music.py file completely and shifted the associated functions to dccommands. Plan to rename references to dccommands to functions soon
