+++ cogs/Dependencies/dccommands.py
@@ -1,21 +1,26 @@
-#dccommands.py defines several different functions necessary for all of the different functions Mobot has
-#It includes several functions that retrieve songs, playlists, Quotes, etc.
+# dccommands.py defines several different functions necessary for all the different functions Mobot has
+# It includes several functions that retrieve songs, playlists, Quotes, etc.
 import asyncio
 import time
 from random import randint
 from nextcord import FFmpegOpusAudio
 import yt_dlp
 from yt_dlp.utils import DownloadError
-import os
 import scrapetube
-import settings
+import os
 import nextcord
+from nextcord import FFmpegPCMAudio
+import validators
+from validators import ValidationFailure
+import os.path
+import settings
+import random
 
-#Sets the working directory
+# Sets the working directory
 pwd = os.path.dirname(os.path.realpath(__file__))
 
-#This function retrieves a techquote from the Funnytechquotes.txt repository of quotes
-#It then returns the quote to be used in Mobot when the command is run
+# This function retrieves a techquote from the Funnytechquotes.txt repository of quotes
+# It then returns the quote to be used in Mobot when the command is run
 def techQuotes():
     files = open(pwd + '/Quotes/' + 'Funnytechquotes.txt', 'r')
     quote = ''
@@ -26,8 +31,9 @@ def techQuotes():
     files.close()
     return quote
 
-#Both SeanQuotes() and oneSeam() retrieve a seanQuote from the repository of Sean Quotes
-#As this is not a function of Mobot, these two functions are useless when this file is included in the Mobot code
+
+# Both SeanQuotes() and oneSeam() retrieve a seanQuote from the repository of Sean Quotes
+# As this is not a function of Mobot, these two functions are useless when this file is included in the Mobot code
 def seanQuotes():
     files = open(pwd + '/Quotes/' + 'Seanquotes.txt', 'r')
     quote = ''
@@ -38,6 +44,7 @@ def seanQuotes():
     files.close()
     return quote
 
+
 def oneSeam():
     files = open(pwd + '/Quotes/' + 'Seanquotes.txt', 'r')
     quote = ''
@@ -46,42 +53,47 @@ def oneSeam():
     files.close()
     return quote
 
-#The loggerOutputs class defines a logger used for yt_dlp
-#This logger is then used to output to STDOUT and STDERR
-#It also generates files with yt_dlp logs for individual servers
+
+# The loggerOutputs class defines a logger used for yt_dlp
+# This logger is then used to output to STDOUT and STDERR
+# It also generates files with yt_dlp logs for individual servers
 class loggerOutputs:
     def __init__(self, ctx):
         self.ctx = ctx
+
     def error(self, msg):
         with open(f'logs/{self.ctx.guild.name}_logs.log', 'a+') as file:
             file.write("Error: " + msg + "\n")
             print("Error: " + msg)
+
     def warning(self, msg):
         with open(f'logs/{self.ctx.guild.name}_logs.log', 'a+') as file:
             file.write("Warning: " + msg + "\n")
             print("Warning: " + msg)
+
     def debug(self, msg):
         with open(f'logs/{self.ctx.guild.name}_logs.log', 'a+') as file:
             file.write("Log: " + msg + "\n")
             print(msg)
 
-#RetrieveAudio defines a function which downloads a youtube video and converts it to an .opus file to be played by Mobot
-async def retrieveAudio(url:str, path:str, ctx, index):
 
-#ydl_ops defines a set of options used to run yt_dlp and get the desired output
+# RetrieveAudio defines a function which downloads a youtube video and converts it to an .opus file to be played by
+# Mobot
+async def retrieveAudio(url: str, path: str, ctx, index):
+    # ydl_ops defines a set of options used to run yt_dlp and get the desired output
     ydl_opts = {
-    'format': 'bestaudio/best',
-    'logger': loggerOutputs(ctx=ctx),
-    'outtmpl': path + '/%(title)s.%(ext)s',
-    'postprocessors': [{
-        'key':'FFmpegExtractAudio',
-        'preferredquality': '0',
-    }],
+        'format': 'bestaudio/best',
+        'logger': loggerOutputs(ctx=ctx),
+        'outtmpl': path + '/%(title)s.%(ext)s',
+        'postprocessors': [{
+            'key': 'FFmpegExtractAudio',
+            'preferredquality': '0',
+        }],
     }
 
-#This then extracts the video from youtube and grabs the necessary information
-#Its all done from the folder for each specific server
-#It then returns the audio source, the title, thumbnail, and duration of the video
+    # This then extracts the video from youtube and grabs the necessary information
+    # Its all done from the folder for each specific server
+    # It then returns the audio source, the title, thumbnail, and duration of the video
     loop = asyncio.get_event_loop()
     with yt_dlp.YoutubeDL(ydl_opts) as ydl:
         try:
@@ -96,21 +108,23 @@ async def retrieveAudio(url:str, path:str, ctx, index):
             print(f"\n{extension}\n")
         except DownloadError:
             print("The Song has failed to Download")
-            channel = nextcord.utils.get(settings.channels[ctx.guild.id].guild.channels, id=settings.channels[ctx.guild.id].channel.id)
+            channel = nextcord.utils.get(settings.channels[ctx.guild.id].guild.channels,
+                                         id=settings.channels[ctx.guild.id].channel.id)
             await channel.send("The current Track has failed to download. The next Track will now Download")
-            return await retrieveAudio(settings.queues[ctx.guild.id][0], (pwd+'/'+str(ctx.guild.id)), ctx, 0)
+            return await retrieveAudio(settings.queues[ctx.guild.id][0], (pwd + '/' + str(ctx.guild.id)), ctx, 0)
     if extension == "webm":
         extension = "opus"
     for file in os.listdir(path):
         if file.endswith(f".{extension}"):
-            os.rename(path+'/'+ file, path+f"/song.{extension}")
-    source = await FFmpegOpusAudio.from_probe(path+f"/song.{extension}")
+            os.rename(path + '/' + file, path + f"/song.{extension}")
+    source = await FFmpegOpusAudio.from_probe(path + f"/song.{extension}")
     times = time.gmtime(info["duration"])
     duration = time.strftime("%H:%M:%S", times)
     return source, title, info["thumbnails"][0]["url"], duration
 
-#This function retrieves a playlist from youtube using scrapetube and pushes the urls to queue
-#It also retrieves the titles of each song and pushes it to queue as well
+
+# This function retrieves a playlist from youtube using scrapetube and pushes the urls to queue
+# It also retrieves the titles of each song and pushes it to queue as well
 async def retrievePlaylist(url, ctx):
     id = url.lstrip('https://www.youtube.com/playlist?list=')
     loop = asyncio.get_event_loop()
@@ -118,10 +132,102 @@ async def retrievePlaylist(url, ctx):
     songlist = []
     title = []
     for video in videos:
-        songlist.append('https://www.youtube.com/watch?v='+ video['videoId'])
+        songlist.append('https://www.youtube.com/watch?v=' + video['videoId'])
         title.append(video['title']['runs'][0]['text'])
-    channel = nextcord.utils.get(settings.channels[ctx.guild.id].guild.channels, id=settings.channels[ctx.guild.id].channel.id)
+    channel = nextcord.utils.get(settings.channels[ctx.guild.id].guild.channels,
+                                 id=settings.channels[ctx.guild.id].channel.id)
     print("Playlist retrieved successfully")
     await channel.send('Playlist Retrieved Successfully')
     return songlist, title
-    
+
+def checkurl(url_string: str):
+    result = validators.url(url_string)
+
+    if isinstance(result, ValidationFailure):
+        return False
+
+    return result
+
+
+# The queue function is what runs the entire music bot.
+# This function is used to periodically check if a song is ready to be loaded up into the voice chat for playing
+async def queue(ctx, client):
+    # First it sets the working directory and checks if the bot is playing a song
+    pwd = os.path.dirname(os.path.realpath(__file__))
+    voice = nextcord.utils.get(client.voice_clients, guild=ctx.guild)
+    if (voice.is_playing() or voice.is_paused()):
+        pass
+    else:
+
+        # It then checks if there is an active queue for an individual server
+        if settings.queues[ctx.guild.id]:
+            if settings.queues[ctx.guild.id][0].startswith('song'):
+                source = FFmpegPCMAudio(pwd + '/' + str(ctx.guild.id) + '/' + settings.queues[ctx.guild.id][0])
+
+            # It then sets up everything for the next song to play properly
+            # It clears the guild directory and sets downloading to true
+            else:
+                settings.downloading[ctx.guild.id][0] = True
+                os.system('rm ' + str(ctx.guild.id) + '/*.opus')
+                os.system('rm ' + str(ctx.guild.id) + '/*.webm')
+
+                # It then checks if shuffle is turned on and grabs the index for the next shuffle
+                if settings.downloading[ctx.guild.id][2] and (not settings.indexes[ctx.guild.id]):
+                    if len(settings.queues[ctx.guild.id]) > 1:
+                        if settings.downloading[ctx.guild.id][1]:
+                            index = random.randint(1, (len(settings.queues[ctx.guild.id]) - 1)) - 1
+                        else:
+                            index = random.randint(1, len(settings.queues[ctx.guild.id])) - 1
+                    else:
+                        index = 0
+                else:
+                    index = 0
+                    settings.indexes[ctx.guild.id] = False
+
+                # It then checks if the next item is a playlist and retrieves every item in the playlist
+                url = settings.queues[ctx.guild.id][index]
+                if "playlist" in url and ("youtube" in url or "youtu.be" in url):
+                    songlist, title = await retrievePlaylist(settings.queues[ctx.guild.id][index], ctx)
+                    voice.stop()
+                    settings.queues[ctx.guild.id].pop(index)
+                    settings.queues[ctx.guild.id] = songlist + settings.queues[ctx.guild.id]
+                    settings.titles[ctx.guild.id].pop(index)
+                    settings.titles[ctx.guild.id] = title + settings.titles[ctx.guild.id]
+
+                # After that it then retrieves the next audio and if it is set to repeating, it places the song back
+                # to the end of the queue It then plays the next song and sets downloading to false
+                else:
+                    if settings.downloading[ctx.guild.id][1]:
+                        settings.titles[ctx.guild.id].append(settings.titles[ctx.guild.id][index])
+                        settings.queues[ctx.guild.id].append(settings.queues[ctx.guild.id][index])
+                    source, title, thumbnail, duration = await retrieveAudio(
+                        settings.queues[ctx.guild.id][index], (pwd + '/' + str(ctx.guild.id)), ctx, index)
+                    textchannel = nextcord.utils.get(settings.channels[ctx.guild.id].guild.channels,
+                                                     id=settings.channels[ctx.guild.id].channel.id)
+                    embed = nextcord.Embed(title="Now playing:", description=title)
+                    embed.set_footer(text=f"Duration: {duration}")
+                    embed.set_thumbnail(url=thumbnail)
+                    await textchannel.send(embed=embed)
+                    # Reminder, ARRAY POPPING FOR TITLES AND QUEUES IS IN dccommands.py
+                    # settings.titles[ctx.guild.id].pop(index)
+                    if settings.downloading[ctx.guild.id][3]:
+                        # loop = asyncio.get_event_loop()
+                        print("normalized")
+                        # raw = await loop.run_in_executor(None, AudioSegment.from_file, f"{pwd}/{ctx.guild.id}/song.opus", codec = "opus")
+                        # raw = AudioSegment.from_file(f"{pwd}/{ctx.guild.id}/song.opus", codec = "opus")
+                        # normalized = effects.normalize(raw, headroom=10)
+                        # normalized = await loop.run_in_executor(None, effects.normalize, raw, headroom=10)
+                        # os.system('rm ' + str(ctx.guild.id) + '/*.opus')
+                        # normalized.export(f"{pwd}/{ctx.guild.id}/song.opus", format="opus")
+                        # await loop.run_in_executor(None, normalized.export, f"{pwd}/{ctx.guild.id}/song.opus", format="opus")
+                        # source = FFmpegOpusAudio(f"{pwd}/{ctx.guild.id}/song.opus")
+                    player = voice.play(source)
+                    # settings.queues[ctx.guild.id].pop(index)
+            settings.downloading[ctx.guild.id][0] = False
+
+        # If there is not an active queue, it cleans up and pauses the timer
+        else:
+            os.system('rm ' + str(ctx.guild.id) + '/*.opus')
+            os.system('rm ' + str(ctx.guild.id) + '/*.webm')
+            await settings.timers[ctx.guild.id].pause()
+            print('No queued items')
